Update enchanted-links.html
The Futuristic Enchantment of the Fairy Alpha King, the link doesn't go to the right page. It needs to go to the following location.

https://skinny-pinks.github.io/my-website/


ChatGPT
Understood. Here is the updated enchanted-links.html with the corrected link for "The Futuristic Enchantment of the Fairy Alpha King":

Updated enchanted-links.html with Corrected Link