Update enchanted-links.html
Okay, that's perfect. Now, I want to add an additional link. I'm going to give you the link, and then a link to where the picture is hosted. I want you to use it like an icon, just like it already exists on this page. Understand?

ChatGPT
Absolutely! I understand. Please provide the link and the link to where the picture is hosted, and I'll add it to your page following the existing style.

link:
https://music.apple.com/profile/skinny_pinks

link to the picture for use as an icon
https://raw.githubusercontent.com/Skinny-Pinks/my-website/main/Link%20to%20my%20Apple%20Music.JPEG

ChatGPT
Got it! I'll add the new link and use the provided picture as an icon. Here is the updated HTML code with the additional link:





Explanation of Additions:
Added a new link box with the provided Apple Music link and picture.
Let me know if you need any further adjustments!